# specs/001-druids-build-a/contracts/api.yaml
openapi: 3.0.3
info:
  title: Druids Multi-Agent System API
  version: 1.0.0
  description: MCP Server interface for managing agents, scenarios, and coordination
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /agents:
    get:
      summary: List all agents
      operationId: listAgents
      parameters:
        - name: realmId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [druid, elemental, gaia, worldtree]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    
    post:
      summary: Create new agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid agent configuration
        '409':
          description: Agent ID already exists
...
components:
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [druid, elemental, gaia, worldtree]
        realmId:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        configuration:
          $ref: '#/components/schemas/AgentConfiguration'
    AgentConfiguration:
      type: object
      properties:
        llmModel:
          type: string
        systemPrompt:
          type: string
        toolAccess:
          type: array
          items:
            $ref: '#/components/schemas/ToolAccessPolicy'
        knowledgeAccess:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeAccessPolicy'
        specialization:
          $ref: '#/components/schemas/SpecializationProfile'
        persona:
          $ref: '#/components/schemas/DruidPersona'
    CreateAgentRequest:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [druid, elemental, gaia, worldtree]
        realmId:
          type: string
        configuration:
          $ref: '#/components/schemas/AgentConfiguration'
      required:
        - id
        - type
        - realmId
        - configuration
...
    DruidPersona:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        coordinationStyle:
          type: string
          enum: [collaborative, directive, consultative, delegative]
        decisionMakingApproach:
          type: string
        managementPrinciples:
          type: array
          items:
            type: string
        elementalBindings:
          type: array
          items:
            type: string
    SpecializationProfile:
      type: object
      properties:
        id:
          type: string
        domain:
          type: string
        expertise:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        defaultPrompts:
          type: array
          items:
            type: string
...
    KnowledgeAccessPolicy:
      type: object
      properties:
        agentId:
          type: string
        namespacePath:
          type: string
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, execute]
    ToolAccessPolicy:
      type: object
      properties:
        agentId:
          type: string
        mcpServerId:
          type: string
        toolName:
          type: string
        permissions:
          type: array
          items:
            type: string
            enum: [invoke, configure]
...